CU-22ck02t ファイルを移動した
