+++ backend/app/Http/Requests/Friends/StoreFriendRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Friends;
 
 use Illuminate\Foundation\Http\FormRequest;
