+++ backend/app/Http/Requests/Friends/UpdateFriendRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Friends;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreResponseRequest extends FormRequest
+class UpdateFriendRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
