+++ backend/app/Http/Requests/Genres/StoreGenreRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Genres;
 
 use Illuminate\Foundation\Http\FormRequest;
