+++ backend/app/Http/Requests/Goods/UpdateGoodRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Goods;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateGenreRequest extends FormRequest
+class UpdateGoodRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
