+++ backend/app/Http/Requests/Groups/StoreGroupRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Groups;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class UpdateGoodRequest extends FormRequest
+class StoreGroupRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
