+++ backend/app/Http/Requests/Responses/StoreResponseRequest.php
@@ -1,10 +1,10 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Responses;
 
 use Illuminate\Foundation\Http\FormRequest;
 
-class StoreGoodRequest extends FormRequest
+class StoreResponseRequest extends FormRequest
 {
     /**
      * Determine if the user is authorized to make this request.
