+++ backend/app/Http/Requests/Responses/UpdateResponseRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Responses;
 
 use Illuminate\Foundation\Http\FormRequest;
